Move delegates to separate file that is ignored for flake8
