+++ tox.ini
@@ -3,7 +3,7 @@ max-line-length = 100
 docstring-convention = all
 import-order-style = pycharm
 print("hey")
-exclude = .idea/,pyinstaller_build/,.venv
+exclude = .idea/,pyinstaller_build/,.venv,auto_neutron/delegates.py
 ignore =
     B311,W503,E226,S311,T000
 # Missing Docstrings
